Injected real weather forecast data into the SPA
I made a function that utilised the data sent by axios from the forecast API and looped through the response for each of the seven days and injected the maximum and minimum temperature of that day into the HTML structure in the displayForecast function.

to limit the days to just 5, I made the function for displayForecast only loop through up to the 5th day in the index (0-4).

Inside of the displayForecast function I also called a function that would format the day to the name of a day of the week instead of an index number. Inside of the function I parsed the timestamp and then assigned the given number a name with the array of days of the week within the formatting function.